Lesson 4, task 1 revision. Refactor code for structure project.
