+++ src/main/java/Lesson1/App.java
@@ -1,4 +1,4 @@
-package Lesson1;
+package main.java.Lesson1;
 
 import java.util.Scanner;
